Separate Weakspot Logic
