+++ Assets/Prefabs/Boss.prefab
@@ -342,7 +342,7 @@ Transform:
   - {fileID: 4876552306627590372}
   - {fileID: 5985919299896830209}
   - {fileID: 367596013244964437}
-  - {fileID: 9048418301526507657}
+  - {fileID: 4990315813721398206}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
@@ -379,22 +379,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d886e5e2aeca74bb1b740eb723715091, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  animator: {fileID: 367596013406301719}
   bossStartColor: {r: 1, g: 0, b: 0, a: 1}
   myColorRed: {r: 0.79607844, g: 0.3254902, b: 0.3254902, a: 1}
-  carrotSpawnOffset: {fileID: 367596013244964437}
-  carrotPf: {fileID: 5649900439809378948, guid: 1eecc96bfe961436396ca41d4ed49458, type: 3}
-  carrotTargetPos: {fileID: 0}
-  carrotDelayTime: 1.6
-  carrotSpeed: 3
   bossTransform: {fileID: 185579822364805374}
   bossHorizontalRange: 3
   bossHorizontalSpeed: 1
   bossSmoke: {fileID: 340498997}
   bossReducedSize: 0.4
-  weakSpotPf: {fileID: 6453732741328588661, guid: a102f0d914b174bf6930588c3c1628fd, type: 3}
-  weakSpotSound: {fileID: 0}
-  weakSpotVolume: 0.7
 --- !u!198 &340498997
 ParticleSystem:
   m_ObjectHideFlags: 0
@@ -10046,65 +10037,6 @@ Transform:
   m_Father: {fileID: 8242001519030258129}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &985909613755199700
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 9048418301526507657}
-  - component: {fileID: 5281379874619515182}
-  - component: {fileID: 8678327652940513097}
-  m_Layer: 7
-  m_Name: BossCombinedMesh
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &9048418301526507657
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 985909613755199700}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 185579822364805374}
-  m_RootOrder: 4
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!64 &5281379874619515182
-MeshCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 985909613755199700}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 4
-  m_Convex: 0
-  m_CookingOptions: 30
-  m_Mesh: {fileID: 2534964839176971238, guid: 9eb69f619f11342129523d2e9d977c14, type: 3}
---- !u!114 &8678327652940513097
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 985909613755199700}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5eeb96102065347958a9f66bbe8f3a0a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1104112380868777928
 GameObject:
   m_ObjectHideFlags: 0
@@ -15118,6 +15050,67 @@ Transform:
   m_Father: {fileID: 1591184725506727710}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &3042987600192495292
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 4990315813721398206}
+  - component: {fileID: 958706580565015589}
+  - component: {fileID: 8514985125646259595}
+  m_Layer: 7
+  m_Name: BossCombinedMesh
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &4990315813721398206
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3042987600192495292}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 185579822364805374}
+  m_RootOrder: 4
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!64 &958706580565015589
+MeshCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3042987600192495292}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_Enabled: 1
+  serializedVersion: 4
+  m_Convex: 0
+  m_CookingOptions: 30
+  m_Mesh: {fileID: 2534964839176971238, guid: 9eb69f619f11342129523d2e9d977c14, type: 3}
+--- !u!114 &8514985125646259595
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3042987600192495292}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5eeb96102065347958a9f66bbe8f3a0a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  weakSpotSound: {fileID: 0}
+  weakSpotVolume: 0.7
 --- !u!1 &3074304981892188758
 GameObject:
   m_ObjectHideFlags: 0
