+++ Assets/Scripts/Boss/BossControl.cs
@@ -25,23 +25,7 @@ public class BossControl : MonoBehaviour
     Dictionary<Transform, Vector3> bossComponentsInitialPositions = new Dictionary<Transform, Vector3>();
     Dictionary<Transform, Quaternion> bossComponentsInitialRotations = new Dictionary<Transform, Quaternion>();
 
-    [Header("Boss weak spot variables")]
-    [SerializeField] GameObject weakSpotPf;
-    Mesh bossMesh;
-    Camera mainCamera;
-    MeshFilter meshFilter;
-    LayerMask occlusionMask;
-    Dictionary<int, Color> WeakSpotStatCol = new Dictionary<int, Color>();
-    Color WeakSpotCols0 = new(51f / 255f, 1f, 0f, 156f / 255f);  // Initial spot color
-    Color WeakSpotCols1 = new(0f, 1f, 219f / 255f, 156f / 255f); // Spot color on first hit
-    Color WeakSpotCols2 = new(0f, 152f / 255f, 1f, 1f);         //      ''       second hit
-    Color WeakSpotCols3 = new(1f, 1f, 1f, 100f / 255f);         //      ''       third hit = final color
-    int hitCount = 0;   // Number of total hits on weakspot
     BossStageManager bossStageManager;
-
-    [Header("Audio Settings")]
-    [SerializeField] public AudioClip weakSpotSound;
-    [SerializeField][Range(0f, 1f)] public float weakSpotVolume = 0.7f;
 
     // Start is called before the first frame update
     void Awake()
@@ -61,29 +45,12 @@ public class BossControl : MonoBehaviour
             bossComponentsInitialRotations[childTransform] = childTransform.rotation;
         }
 
-        // Weak spot
-        mainCamera = Camera.main;
-        meshFilter = gameObject.GetComponent<MeshFilter>();
-        bossMesh = meshFilter.mesh;
-        occlusionMask = gameObject.layer;
-        WeakSpotStatCol = new Dictionary<int, Color>()
-        {
-            { 0, WeakSpotCols0 },
-            { 1, WeakSpotCols1 },
-            { 2, WeakSpotCols2 },
-            { 3, WeakSpotCols3 },
-        };
-
         bossStageManager = GameObject.Find("BossStageManager").GetComponent<BossStageManager>();
-
-        //BossDeath(); // Used to test boss death in scene
     }
 
     void Start()
     {
         playerTransform = bossStageManager.ActiveCharacter().transform;
-
-        GetWeakSpots();
     }
 
     // Update is called once per frame
@@ -136,7 +103,6 @@ public class BossControl : MonoBehaviour
         if (!bossDead)
         {
             bossDead = true;
-            RemoveAllWeakSpots();
 
             // Fall back effect
             foreach (Transform childTransform in bossTransform)
@@ -188,179 +154,6 @@ public class BossControl : MonoBehaviour
         bossTransform.localScale *= bossReducedSize;
     }
 
-    // Create weak spots : Call at the start of each phase
-    void GetWeakSpots()
-    {
-        // Get all vertices on boss mesh
-        Vector3[] vertices = bossMesh.vertices;
-
-        // Divide target region into 8 areas
-        Bounds bounds = bossMesh.bounds;
-        Vector3 center = meshFilter.transform.TransformPoint(bounds.center);
-
-        float quarterHeight = bounds.size.y / 4;
-        float lowerY = center.y - quarterHeight;
-        float middleY = center.y;
-        float upperY = center.y + quarterHeight;
-
-        List<Vector3>[] regions = new List<Vector3>[8];
-        for (int i = 0; i < 8; i++) regions[i] = new List<Vector3>();
-
-        // Reduce vertices
-        foreach (Vector3 vertex in vertices)
-        {
-            Vector3 worldPos = meshFilter.transform.TransformPoint(vertex);
-            Vector3 viewportPos = mainCamera.WorldToViewportPoint(worldPos);
-
-            // Check if vertex is within the camera’s frustum
-            if (viewportPos.x < 0 || viewportPos.x > 1 || viewportPos.y < 0 || viewportPos.y > 1 || viewportPos.z <= 0) continue;
-
-            // Perform an occlusion check
-            if (Physics.Linecast(mainCamera.transform.position, worldPos, occlusionMask)) continue;
-
-            int index = 0;
-
-            // X-axis split: left or right
-            if (worldPos.x > center.x) index += 1;
-
-            // Y-axis split: bottom, lower middle, upper middle, or top
-            if (worldPos.y > upperY) index += 6;        // Top region
-            else if (worldPos.y > middleY) index += 4;  // Upper middle region
-            else if (worldPos.y > lowerY) index += 2;   // Lower middle region
-            // Bottom region does not need an additional offset
-
-            //|----|----|
-            //|  6 | 7  |
-            //|----|----| upper y
-            //| 4  |  5 |
-            //|----|----| middle y
-            //|  2 |  3 |
-            //|----|----| lower y
-            //| 0  |  1 |
-            //|----|----|
-
-            regions[index].Add(worldPos);
-
-        }
-
-        // Choose three random regions and select one random point from each
-        List<int> chosenIndices = new List<int>();
-
-        while (chosenIndices.Count < 3)
-        {
-            int randomIndex = Random.Range(0, 8);
-
-            if (!chosenIndices.Contains(randomIndex) && regions[randomIndex].Count > 0)
-            {
-                chosenIndices.Add(randomIndex);
-                Vector3 randomPoint = regions[randomIndex][Random.Range(0, regions[randomIndex].Count)];
-
-                // Make weak spot orientation match boss mesh
-                Vector3 localPoint = meshFilter.transform.InverseTransformPoint(randomPoint);
-                int closestVertexIndex = FindClosestVertexIndex(localPoint);
-
-                Vector3 normal = bossMesh.normals[closestVertexIndex];
-                Vector3 worldNormal = meshFilter.transform.TransformDirection(normal);
-
-                Instantiate(weakSpotPf, randomPoint, Quaternion.LookRotation(worldNormal), gameObject.transform);
-            }
-        }
-    }
-
-    // To find normal (Called by GetWeakSpots())
-    int FindClosestVertexIndex(Vector3 point)
-    {
-        Vector3[] vertices = bossMesh.vertices;
-        int closestIndex = 0;
-        float minDistance = float.MaxValue;
-
-        for (int i = 0; i < vertices.Length; i++)
-        {
-            float distance = Vector3.Distance(vertices[i], point);
-            if (distance < minDistance)
-            {
-                minDistance = distance;
-                closestIndex = i;
-            }
-        }
-
-        return closestIndex;
-    }
-
-    // Call to change color of weak spot on attack (collision between laser and weakspot)
-    public void TransformWeakSpot(GameObject weakSpot)
-    {
-        SpriteRenderer sr = weakSpot.GetComponent<SpriteRenderer>();
-        if (sr == null) return;
-        Color WeakSpotCol = sr.color;
-
-        int status = -1;
-
-        foreach (KeyValuePair<int, Color> statCol in WeakSpotStatCol)
-        {
-            if (statCol.Value == WeakSpotCol)
-            {
-                status = statCol.Key;
-                break;
-            }
-        }
-
-        if (status == 3 || status == -1) return;
-        GameManager.inst.AddScore(3000);
-        if (weakSpotSound != null)
-        {
-            AudioSource.PlayClipAtPoint(weakSpotSound, gameObject.transform.position, weakSpotVolume);
-        }
-        hitCount += 1;
-        if (hitCount % 9 == 0)
-        {
-            bossStageManager.IncrementPhase();
-            if (bossStageManager.GetPhase() < 3) Invoke("NewWeakSpots", 0.2f);
-        }
-        StartCoroutine(gradualColorChange(sr, sr.color, WeakSpotStatCol[status + 1]));
-    }
-
-    // Called by TransformWeakSpot()
-    IEnumerator gradualColorChange(SpriteRenderer sr, Color startCol, Color endCol)
-    {
-        float elapsedTime = 0f;
-        float duration = 0.1f;
-
-        while (elapsedTime < duration)
-        {
-            elapsedTime += Time.deltaTime;
-            if (sr != null)
-                sr.color = Color.Lerp(startCol, endCol, elapsedTime / duration);
-
-
-            yield return null;
-        }
-
-        if (sr != null) sr.color = endCol;
-    }
-
-    // Called on boss death, call on phase change
-    void RemoveAllWeakSpots()
-    {
-        GameObject[] wss = GameObject.FindGameObjectsWithTag("WeakSpot");
-        foreach (GameObject ws in wss) Destroy(ws);
-    }
-
-
-    // Testing only. Removes old and produces new weakspots.
-    public void NewWeakSpots()
-    {
-        RemoveAllWeakSpots();
-        GetWeakSpots();
-    }
-
-    //// Testing only. Changes color of weakspots one by one.
-    //public void TransformWeakSpotHelper()
-    //{
-    //    foreach (Transform child in transform)
-    //        if (child.gameObject.CompareTag("WeakSpot"))
-    //            if (TransformWeakSpot(child.gameObject)) return;
-    //}
     public bool IsDead()
     {
         return bossDead;
