+++ Assets/Scripts/Boss/Weakspot/TransformWeakspot.cs.meta
