+++ Assets/Scripts/Boss/Weakspot/WeakspotManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b90d792ddfabd42bcbb39d5370814918
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
