+++ ProjectSettings/TagManager.asset
@@ -16,6 +16,7 @@ TagManager:
   - WeakSpot
   - Fire
   - Laser
+  - BossMesh
   layers:
   - Default
   - TransparentFX
@@ -29,6 +30,26 @@ TagManager:
   - Player
   - DropObjectMesh
   - DropObjectBox
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
+  - 
   m_SortingLayers:
   - name: Default
     uniqueID: 0
